requirements.txt
requests~=2.31.0
urllib3~=2.0.4
jsonschema~=4.18.6
python-dotenv~=1.0.0
transliterate~=1.10.2
matplotlib~=3.7.2